app-text/requirements.txt
matplotlib==3.8.4
networkx==3.2.1
networkx==3.3
nltk==3.8.1
pyvis==0.3.1
sematch.egg==info
sense2vec==2.0.2
stanford_openie==1.3.2
streamlit==1.32.0
